I init getCreators lib to keep API endpoint clean.g
